Add go.mod file
This PR makes sure to add a `go.mod` file, which will help to distribute
`go-mysqlstack` in an easier way going forward. I think we should also
tag the library with a version, such as `v1.0.0`. That will help others
to pick up a version in the future if we release newer versions of this
library.